requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=3.14
psycopg2-binary>=2.9.9
Pillow>=9.0
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
sentence-transformers
requests
gunicorn
whitenoise

# ML Server Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
torch>=2.2.0
transformers>=4.41.0
python-multipart
numpy<2.0
soundfile
librosa